build.sh: build kernel in subdir of source tree
As it turned out, for this kernel version, building in-tree and
out-of-tree means different cflags. Specifically, building out-of-tree
adds $(obj) to list of include dirs due to this line in
scripts/Makefile.lib:

__c_flags	= $(call addtree,-I$(obj)) $(call flags,_c_flags)

Core kernel code doesn't care, but mediatek code inside
drivers/misc/mediatek depends on this behaviour, so breaks when building
in-tree due to missing -I options in cflags. build.sh accounts for that
and builds in out/ subdir of source tree.

This is how old kernels are built by pmOS, which requires setting _outdir
var in APKBUILD. The dir itself is created by downstreamkernel_prepare
script from devicepkg-dev package.